drm/amdgpu: fix compile warning in amdgpu_kms.c
warning:
comparison of distinct pointer types lacks a cast
 min((size_t)size, bios_size - bios_offset))

+++ drivers/gpu/drm/amd/amdgpu/amdgpu_kms.c
@@ -561,7 +561,7 @@ static int amdgpu_info_ioctl(struct drm_device *dev, void *data, struct drm_file
 
 			bios = adev->bios + bios_offset;
 			return copy_to_user(out, bios,
-				min((size_t)size, bios_size - bios_offset))
+				min(size, bios_size - bios_offset))
 					? -EFAULT : 0;
 		}
 		default:
